requirements.txt
numpy
pandas
hydra-core
mlflow
scikit-learn
pytest
requests
fastapi==0.63.0
uvicorn
gunicorn
dvc[s3]